Add an abstract class for GameOfLife.cs.
The new abstract class BaseGameOfLife.cs reduces coupling and enhances the maintenance and extensibility of the system.
